src/channel_web/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.27.2
jinja2==3.1.4
python-multipart==0.0.9
google-auth==2.35.0
requests==2.32.3
